Use cmake build to solve errors a W3
Someday we want to enable this in the gradle build, but it will take a bit.

Cmake helps, and this fixes wpiutil, ntcore and the simulator hal

Working on wpilibc and cscore, need opencv working in cmake on windows first

+++ hal/src/main/native/include/mockdata/SimCallbackRegistry.h
@@ -30,7 +30,7 @@ class SimCallbackRegistryBase {
  public:
   void Cancel(int32_t uid) {
     std::lock_guard<wpi::recursive_spinlock> lock(m_mutex);
-    if (m_callbacks) m_callbacks->erase(uid - 1);
+    if (m_callbacks) m_callbacks->erase(static_cast<size_t>(uid) - 1);
   }
 
   void Reset() {
